Consolidate color variables
The original code used shades that were extremely close to each other for the different Baseline statuses: one for the Baseline checkmark logo, another for the status icons shown next to browsers. The difference seemed more like a mistake than anything and made the color API surface bigger than necessary, so this commit merges those to use the same shades for both.